backend/requirements.txt
fastapi
uvicorn
pymupdf
sentence-transformers
pymongo
python-dotenv
python-multipart
requests
torch
transformers
